Переносит логику отрисовки в tripController
